first complete iteration of edit/delete

+++ routes/tweet_routes.py
@@ -54,7 +54,7 @@ def get_tweets(page):
             "comment_count": len(comments)
         })
 
-    return render_template("tweetsPage.html", tweets=tweet_data, nextPage=page + 1)
+    return render_template("tweetsPage.html", tweets=tweet_data, nextPage=page + 1, user=user)
 
 
 @bp.post("/tweets/<int:tweet_id>/like")
@@ -122,15 +122,16 @@ def create_comment(tweet_id):
 
     return render_template("commentSection.html", comments=comments, tweet=tweet, user=user)
 
-@bp.post("/delete")
-def delete_message():
-    message = request.form.get('message')
+@bp.delete("/delete/<int:tweet_id>")
+def delete_message(tweet_id):
     userID = request.cookies.get('userId')
     if not userID:
-        return redirect('/')
+        return "<script>window.location = '/'</script>"
 
-    user = User.find(user_id=userID)
-    tweet = Tweet.select().where(Tweet.message == message, Tweet.user == user).first()
+    tweet = Tweet.find(tweet_id)
+
+    if tweet.user.userID != userID:
+        return "<script>window.location = '/'</script>"
 
     if tweet:
         tweet.delete_instance()
@@ -139,4 +140,30 @@ def delete_message():
 
     tweet.save()
 
-    return render_template("tweet.html")
+    return ""
+
+
+@bp.get("/tweets/editMessage/<editing>/<int:tweet_id>")
+def edit_message(editing, tweet_id):
+    userID = request.cookies.get('userId')
+    if not userID:
+        return "<script>window.location = '/'</script>"
+    tweet = Tweet.find(tweet_id)
+
+    if tweet.user.userID != userID:
+        return "<script>window.location = '/'</script>"
+
+    if editing == "True":
+        return render_template("editMessage.html", tweet=tweet)
+
+
+@bp.post("/tweets/<int:tweet_id>/update/")
+def update_tweet(tweet_id):
+    userID = request.cookies.get('userId')
+    if not userID:
+        return "<script>window.location = '/'</script>"
+
+    tweet = Tweet.find(tweet_id)
+    tweet.message = request.form['message']
+    tweet.save()
+    return f"<p id='tweet-text-{ tweet.id }'style='font-size: 18px; line-height: 1.6; font-weight: bold; color: #f0f1f3;'> {tweet.message} </p>"
